feat(web): convert /dashboard and /score/[domain] to hono/jsx
Two more page routes migrated, including the most complex one:
- dashboard: Layout component wraps each view; inline login <script> via raw()
  (preserves => arrows); owner-isolation rendering unchanged (tests confirm no
  cross-owner leakage). page() now stringifies a <Layout> tree.
- score/[domain]: the score hub — inline SVG ChartSvg + AxisChip components,
  dynamic per-tier CSS (c.fg interpolated), conditional claim/dofollow + peer +
  chart + tells blocks, and the share/claim inline <script> (with embedded
  JSON.stringify values) via raw(). Empty/404 state is its own JSX doc.

Verified: dashboard.test.js (13/13), score.test.js (7/7), full web suite 122/122,
tsc clean, wrangler pages functions build compiles the Worker.
